+++ gems/capybara-dommy/spec/unit/configuration_spec.rb
@@ -11,6 +11,23 @@ RSpec.describe Capybara::Dommy::Configuration do
     expect(config.follow_redirects).to be(true)
     expect(config.max_redirects).to eq(5)
     expect(config.visibility).to eq(:html)
+    expect(config.raise_js_errors).to be(true)
+  end
+
+  it "passes raise_js_errors to a javascript driver's session" do
+    driver = Capybara::Dommy::Driver.new(proc {}, javascript: true)
+    expect(driver.send(:instance_variable_get, :@session_options)[:strict_js_errors]).to be(true)
+  end
+
+  it "leaves a non-javascript driver's session alone" do
+    driver = Capybara::Dommy::Driver.new(proc {})
+    expect(driver.send(:instance_variable_get, :@session_options)).not_to have_key(:strict_js_errors)
+  end
+
+  it "can turn JS-error failures off process-wide" do
+    Capybara::Dommy.configure { |config| config.raise_js_errors = false }
+    driver = Capybara::Dommy::Driver.new(proc {}, javascript: true)
+    expect(driver.send(:instance_variable_get, :@session_options)[:strict_js_errors]).to be(false)
   end
 
   it "sets values via the configure block" do